Add watchdog that terminates script on timeout.
Had some issues with ECHONET Lite polls hanging indefinitely without
ever timing out. Figured a general watchdog across the entire app is the
best solution, since this also provides a solution for when discovery
has failed for some reason and there are no devices to poll.

With the app simply terminating, the underlying process manager can
restart it as appropiate.